refs #105 Refactor compose component
