+++ src/pages/index.tsx
@@ -12,7 +12,7 @@ import NewTimeline from 'src/components/timelines/New'
 import ShowTimeline from 'src/components/timelines/Show'
 import NewServer from 'src/components/servers/New'
 import Navigator from 'src/components/Navigator'
-import Compose from 'src/components/Compose'
+import Compose from 'src/components/compose/Compose'
 import Media from 'src/components/Media'
 import generateNotification from 'src/utils/notification'
 import { ReceiveNotificationPayload } from 'src/payload'
